# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PMS Hotel Management System API
  description: |
    Complete REST API documentation for the Property Management System.
    
    ## Authentication
    All endpoints (except login) require a Bearer token in the Authorization header.
    
    ## Base URL
    - Development: http://localhost:3001/api
    - Production: https://api.yourhotel.com/api
    
    ## Error Codes
    - 400 - Bad Request
    - 401 - Unauthorized
    - 403 - Forbidden
    - 404 - Not Found
    - 500 - Internal Server Error
  version: 1.0.0
  contact:
    name: API Support
    email: support@grandview.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.grandview.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /invoices:
    get:
      tags:
        - Invoices
      summary: Get all invoices
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: reservationId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of invoices
    post:
      tags:
        - Invoices
      summary: Create new invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reservationId:
                  type: integer
                invoiceDate:
                  type: string
                  format: date
                dueDate:
                  type: string
                  format: date
                subTotal:
                  type: number
                taxAmount:
                  type: number
                discountAmount:
                  type: number
                totalAmount:
                  type: number
                notes:
                  type: string
                items:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Invoice created
  /invoices/{id}:
    get:
      tags:
        - Invoices
      summary: Get invoice by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Invoice details
    put:
      tags:
        - Invoices
      summary: Update invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invoiceDate:
                  type: string
                dueDate:
                  type: string
                subTotal:
                  type: number
                taxAmount:
                  type: number
                totalAmount:
                  type: number
                status:
                  type: string
      responses:
        '200':
          description: Invoice updated
    delete:
      tags:
        - Invoices
      summary: Cancel invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Invoice cancelled
  /invoices/{id}/items:
    post:
      tags:
        - Invoices
      summary: Add invoice item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                itemType:
                  type: string
                description:
                  type: string
                quantity:
                  type: integer
                unitPrice:
                  type: number
                amount:
                  type: number
                taxRate:
                  type: number
      responses:
        '201':
          description: Item added
  /invoices/{id}/print:
    get:
      tags:
        - Invoices
      summary: Get invoice for printing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Invoice print data

  # ============================================================
  # PAYMENTS
  # ============================================================
...
  /maintenance/requests:
    get:
      tags:
        - Maintenance
      summary: Get maintenance requests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: roomId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of requests
    post:
      tags:
        - Maintenance
      summary: Create maintenance request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roomId
                - requestType
                - description
              properties:
                roomId:
                  type: integer
                requestType:
                  type: string
                priority:
                  type: string
                description:
                  type: string
                reportedBy:
                  type: integer
                scheduledDate:
                  type: string
                estimatedCost:
                  type: number
      responses:
        '201':
          description: Request created
  /maintenance/requests/{id}:
    get:
      tags:
        - Maintenance
      summary: Get maintenance request by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Request details
    put:
      tags:
        - Maintenance
      summary: Update maintenance request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                scheduledDate:
                  type: string
                estimatedCost:
                  type: number
                actualCost:
                  type: number
                notes:
                  type: string
      responses:
        '200':
          description: Request updated
  /maintenance/tasks:
    post:
      tags:
        - Maintenance
      summary: Create maintenance task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - requestId
              properties:
                requestId:
                  type: integer
                assignedStaffId:
                  type: integer
                startDate:
                  type: string
      responses:
        '201':
          description: Task created
  /maintenance/tasks/{id}:
    put:
      tags:
        - Maintenance
      summary: Update maintenance task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                completionDate:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Task updated
  /maintenance/history:
    get:
      tags:
        - Maintenance
      summary: Get maintenance history
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        '200':
          description: History list

  # ============================================================
  # INVENTORY
  # ============================================================
...